File saving

+++ ImageDatasetBuilder/CSVService.cs
@@ -46,10 +46,10 @@ namespace ImageDatasetBuilder
             public double Height { get; set; }
             public double Width { get; set; }
             public string Class { get; set; }
-            public double XMin { get; set; }
-            public double XMax { get; set; }
-            public double YMin { get; set; }
-            public double YMax { get; set; }
+            public int XMin { get; set; }
+            public int XMax { get; set; }
+            public int YMin { get; set; }
+            public int YMax { get; set; }
             
         }
